+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Coreference inference: gender, number and animacy.
 - Coreference inference:'X is Y' relationships.
 - Function meaning.concepts.has_common_or_proper_concept(edge).
+- Function hypergraph.add() now has optional 'count' parameter.
 
 ### Changed
 - parse() now returns a dictionary that includes inferred edges.